[TASK]: OpenCL supported matching string on node_binary_vec ,#P-00003
Detail:
* Simple  Global items No.1024 and local items have No.32.
* Check write back completed matching with node_symbol_vec and
* node_state_vec.
* Error on multithread not improved.
* Plan to use TBB multithread check infected string.

ReviewedBy: R.Chatsiri
AutomergeTo:

Resolves: #
Documentation: #
Release:

	new file:   src/data_structure/ac_multithread_tbb.hpp
	new file:   src/database/hnmav_main.rules
	modified:   src/filetypes/pe_file_controller.cpp
	modified:   src/ocl/cl/tire_parallel.cl
	new file:   src/ocl/cl/tire_parallel.cl~
	modified:   src/ocl/cl_bootstrap.cpp
	modified:   src/ocl/cl_bootstrap.hpp
	modified:   src/ocl/utils/base_clutil.hpp
	modified:   src/ocl/utils/clutil_commandqueue.cpp
	modified:   src/ocl/utils/clutil_memory.cpp
	modified:   src/ocl/utils/clutil_memory.hpp
	modified:   src/scan/file_scan_policy.cpp
	modified:   src/test/Makefile.am
	new file:   src/test/environment_test.hpp
	modified:   src/test/file_policy_test.hpp
	modified:   src/test/semaphore_controller_test.cpp
	modified:   src/test/test_suite
	modified:   src/threadsyncocl/thread_controller.cpp
	modified:   src/threadsyncocl/thread_sync.cpp

+++ src/ocl/cl_bootstrap.cpp
@@ -154,7 +154,7 @@ namespace hnmav_kernel
              >
     bool cl_load_system<UtilPlatform, TireDefine, WorkTypes, ContainerT>::
     cl_process_buffer(std::vector<char>& symbol_vec_ptr,
-            std::vector<size_t>&         state_vec_ptr,
+            std::vector<int>&            state_vec_ptr,
             std::vector<char>&           binary_vec,
             std::vector<uint8_t>&       result_vec)
     {
